# backend/src/TradeCoveApp/api.docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "TradeCOVE API",
        "description": "API for TradeCOVE mobile app",
        "version": "1.0.0"
    },
    "host": "trade.mgbeta.ru",
    "schemes": [
        "https"
    ],
    "basePath": "/api/v1",
    "produces": [
        "application/json"
    ],
    "paths": {
        "/user": {
            "get": {
                "summary": "user api",
                "description": "Позволяет получить список всех юзеров в системе\n",
                "tags": [
                    "user"
                ],
                "responses": {
                    "200": {
                        "description": "Массив пользователей",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Создать пользователя",
                "description": "Создать пользователя\n",
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "type": "string",
                        "description": "username",
                        "required": true
                    },
                    {
                        "name": "firstName",
                        "in": "query",
                        "type": "string",
                        "description": "Имя",
                        "required": true
                    },
                    {
                        "name": "lastName",
                        "in": "query",
                        "type": "string",
                        "description": "Фамилия",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "type": "string",
                        "description": "Email",
                        "required": true
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "type": "string",
                        "description": "Пол",
                        "required": true,
                        "enum": [
                            "male",
                            "female"
                        ]
                    },
                    {
                        "name": "about",
                        "in": "query",
                        "type": "string",
                        "description": "описание пользователя",
                        "required": false
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "type": "number",
                        "description": "рейтинг",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Созданный пользователь",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "type": "string",
                    "description": "ID пользователя",
                    "required": true
                }
            ],
            "get": {
                "summary": "user api",
                "description": "Позволяет найти пользователя по ID\n",
                "tags": [
                    "user"
                ],
                "responses": {
                    "200": {
                        "description": "USER",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            },
            "put": {
                "summary": "Изменить пользователя",
                "description": "Изменить пользователя\n",
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "type": "string",
                        "description": "username",
                        "required": true
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "type": "string",
                        "description": "username",
                        "required": false
                    },
                    {
                        "name": "firstName",
                        "in": "query",
                        "type": "string",
                        "description": "Имя",
                        "required": false
                    },
                    {
                        "name": "lastName",
                        "in": "query",
                        "type": "string",
                        "description": "Фамилия",
                        "required": false
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "type": "string",
                        "description": "Email",
                        "required": false
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "type": "string",
                        "description": "Пол",
                        "required": false,
                        "enum": [
                            "male",
                            "female"
                        ]
                    },
                    {
                        "name": "about",
                        "in": "query",
                        "type": "string",
                        "description": "описание пользователя",
                        "required": false
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "type": "number",
                        "description": "рейтинг",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Созданный пользователь",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/user/{id}/followers": {
            "get": {
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "type": "string",
                        "description": "ID пользователя",
                        "required": true
                    }
                ],
                "summary": "Найти подписчиков пользоватея",
                "description": "Найти подписчиков пользоватея\n",
                "tags": [
                    "user"
                ],
                "responses": {
                    "200": {
                        "description": "USER",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/user/{id}/friends": {
            "get": {
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "type": "string",
                        "description": "ID пользователя",
                        "required": true
                    }
                ],
                "summary": "Найти друзей пользователя",
                "description": "Найти друзей пользоватея\n",
                "tags": [
                    "user"
                ],
                "responses": {
                    "200": {
                        "description": "USER",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/user/{id}/offers": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "type": "string",
                    "description": "ID пользователя",
                    "required": true
                }
            ],
            "get": {
                "summary": "Найти товары пользователя",
                "description": "Найти товары пользователя\n",
                "tags": [
                    "user"
                ],
                "responses": {
                    "200": {
                        "description": "USER",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Offer"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/user/{id}/messages": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "type": "string",
                    "description": "ID пользователя",
                    "required": true
                }
            ],
            "get": {
                "summary": "Найти сообщения пользователя",
                "description": "Найти сообщения пользователя\n",
                "tags": [
                    "user"
                ],
                "responses": {
                    "200": {
                        "description": "USER",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Message"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/interest": {
            "get": {
                "summary": "Массив всех интересов",
                "description": "Массив всех интересов\n",
                "tags": [
                    "interest"
                ],
                "responses": {
                    "200": {
                        "description": "interests",
                        "schema": {
                            "$ref": "#/definitions/Interest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Error"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Позволяет создать новый интерес",
                "description": "Позволяет создать новый интерес\n",
                "tags": [
                    "interest"
                ],
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "type": "string",
                        "description": "Название",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Созданный интерес",
                        "schema": {
                            "$ref": "#/definitions/Interest"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/interest/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "type": "string",
                    "description": "ID интереса",
                    "required": true
                }
            ],
            "get": {
                "summary": "Интерес по ID",
                "description": "Позволяет получить интерес по его ID\n",
                "tags": [
                    "interest"
                ],
                "responses": {
                    "200": {
                        "description": "Интерес",
                        "schema": {
                            "$ref": "#/definitions/Interest"
                        }
                    },
                    "404": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Изменить интерес",
                "description": "Изменить интерес\n",
                "tags": [
                    "interest"
                ],
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "type": "string",
                        "description": "название интереса",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Интерес",
                        "schema": {
                            "$ref": "#/definitions/Interest"
                        }
                    },
                    "404": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/category": {
            "get": {
                "summary": "Категории товаров",
                "description": "Позволяет получить список всех возможных категорий для товаров\n",
                "tags": [
                    "category"
                ],
                "responses": {
                    "200": {
                        "description": "Массив всех категорий",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Category"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Создать новую категорию товаров",
                "description": "Создать новую категорию товаров\n",
                "tags": [
                    "category"
                ],
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "type": "string",
                        "description": "Название категории",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Категория",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/category/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "type": "string",
                    "description": "ID категории",
                    "required": true
                }
            ],
            "get": {
                "summary": "Интерес по ID",
                "description": "Позволяет получить категорию по ID\n",
                "tags": [
                    "category"
                ],
                "responses": {
                    "200": {
                        "description": "Категория",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    },
                    "404": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Изменить категорию товаров",
                "description": "Изменить категорию товаров\n",
                "tags": [
                    "category"
                ],
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "type": "string",
                        "description": "Название категории",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Категория",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/offer": {
            "get": {
                "summary": "Массив товаров",
                "description": "Позволяет получить весь массив товаров\n",
                "tags": [
                    "offer"
                ],
                "responses": {
                    "200": {
                        "description": "Массив всех товаров",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Offer"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Добавить товар",
                "description": "Добавить товар\n",
                "tags": [
                    "offer"
                ],
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "type": "string",
                        "description": "ID юзера, от имени которого будет выставлен товар",
                        "required": true
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "type": "string",
                        "description": "Название товара",
                        "required": true
                    },
                    {
                        "name": "categoryId",
                        "in": "query",
                        "type": "string",
                        "description": "ID категории, к которой относится товар",
                        "required": true
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "type": "string",
                        "description": "описание товара",
                        "required": false
                    },
                    {
                        "name": "condition",
                        "in": "query",
                        "type": "string",
                        "description": "состояние товара. New/Used",
                        "required": false
                    },
                    {
                        "name": "images",
                        "in": "query",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "base64"
                        },
                        "description": "состояние товара. New/Used",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Добавленный товар",
                        "schema": {
                            "$ref": "#/definitions/Offer"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
        "/offer/category/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "type": "string",
                    "description": "ID категории",
                    "required": true
                }
            ],
            "get": {
                "summary": "Найти товары по категории",
                "description": "Найти ктовары по категории\n",
                "tags": [
                    "offer"
                ],
                "responses": {
                    "200": {
                        "description": "Товары конкретной категории",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Offer"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
        "/message": {
            "get": {
                "summary": "Найти все сообщения",
                "description": "Найти все сообщения\n",
                "tags": [
                    "message"
                ],
                "responses": {
                    "200": {
                        "description": "Массив всех сообщений",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Message"
                            }
                        }
                    },
                    "default": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Offer": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "description": "ID"
                },
                "userId": {
                    "type": "string",
                    "description": "ID пользователя, который создал"
                },
                "categoryId": {
                    "type": "string",
                    "description": "ID категории, к которой относится товар"
                },
                "title": {
                    "type": "string",
                    "description": "Название"
                },
                "description": {
                    "type": "string",
                    "description": "Описание товара"
                },
                "condition": {
                    "type": "string",
                    "description": "Состояние товара.",
                    "enum": [
                        "New",
                        "Used"
                    ]
                },
                "createdAt": {
                    "type": "string",
                    "format": "date",
                    "description": "Дата создания"
                },
                "updatedAt": {
                    "type": "string",
                    "format": "date",
                    "description": "Дата последнего изменения"
                }
            }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}